ci: run the full test suite on Windows
The Windows job ran `go test ./internal/flows/...` only. It now runs
`go test ./...`, so a Windows regression in any package fails the pull
request. `test/proxye2e` is in that set. `test/acceptance` is not, because
it sits behind the `acceptance` build tag.

Tests that encode a Unix fact move behind a `unix` build tag or a guard:

- `internal/shim/path_test.go` and `system_test.go`: every table is a
  `:`-joined Unix PATH, and system install is Linux only.
- `internal/shim/shim_test.go`: the executable bit does not exist on
  Windows. Two tests that assert a POSIX shim body move to `shim_unix_test.go`.
  The rc-file assertions compare against `binDir` instead of a `/` literal.
- `internal/doctor/checks_test.go`: `LookPath` on Windows needs a `PATHEXT`
  extension. The venv and `#!/bin/sh` stub tests move to `checks_unix_test.go`.
- `cmd/setup/doctor_test.go`: `os.Chmod` cannot make a directory unwritable
  on Windows.
- `internal/runner/childexit_test.go`: two cases drive `sh` and a signal.
