+++ internal/shim/path_unix_test.go
@@ -1,3 +1,5 @@
+//go:build unix
+
 package shim
 
 import (
